+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 * **CHANGED:** Deprecate the use of an array of chars for string literals; use `str` instead.
 * **CHANGED:** Implement `getelement` instruction for `str` and pointers.
 * **CHANGED:** Implement `ref` instruction.
+* **ADDED:** `builder` module to generate Foxil code files from V
 
 ## [v0.2.0] - 2021-07-26